fix(shared,web): split quota module into client-safe types and server-only DB helpers
