+++ shared/package.json
@@ -13,6 +13,7 @@
     "./platforms/reddit": "./src/platforms/reddit/index.ts",
     "./platforms/reply-reader": "./src/platforms/base-reply-reader.ts",
     "./quota": "./src/quota.ts",
+    "./quota-types": "./src/quota-types.ts",
     "./agents": "./src/agents/base.ts",
     "./agents/claude-code": "./src/agents/claude-code.ts",
     "./agents/registry": "./src/agents/registry.ts",
